refactor(tests): remove ModuleInitializer, fail-fast on missing RIMWORLD_DIR, inline fixture, strip ADR refs
F1: Delete [ModuleInitializer] and net472 ModuleInitializerAttribute polyfill from
RimWorldResolverSetup — [OneTimeSetUp] alone passes all 166 tests. Remove the
AppDomain.GetData/SetData idempotency guard that was only justified by the dual-path
coexistence.

F2: In [OneTimeSetUp] RegisterRimWorldResolver(), validate RIMWORLD_DIR / RimWorldDir is
set and the Managed directory exists before registering the AssemblyResolve handler.
Throw an actionable InvalidOperationException at setup time instead of falling back
to a hard-coded machine-specific path.

F3: Strip all ADR-NNNN / AC-N citations from code comments in RimWorldResolverSetup.cs,
LordKuper.Common.csproj, Directory.Build.props, EnumHelperTests.cs, and MathHelperTests.cs;
replace with plain rationale prose.

F4: Inline StaticStateFixture snapshot/restore logic directly into StaticStateTestBase
[SetUp]/[TearDown]; delete StaticStateFixture.cs (the class had exactly one consumer).
Preserved save/restore set verbatim: DefProvider.Current, StatHelper, WorkTypeStatMap,
SkillStatMap._map, PassionHelper statics, StatRanges.Ranges.
