+++ Source/LordKuper.Common.Tests/RimWorldResolverSetup.cs
@@ -1,37 +1,12 @@
 using System.Reflection;
-using System.Runtime.CompilerServices;
-
-// net472 polyfill: [ModuleInitializer] is a C# 9 / .NET 5+ built-in. The compiler only needs
-// the attribute to exist in the right namespace; the CLR honours it on net472 as long as the
-// compiler emits the .cctor wiring. This stub is internal, compile-only, and invisible to callers.
-namespace System.Runtime.CompilerServices
-{
-    [AttributeUsage(AttributeTargets.Method)]
-    internal sealed class ModuleInitializerAttribute : Attribute { }
-}
 
 // Global (namespace-less) NUnit SetUpFixture so this fixture applies to the whole assembly.
 // NUnit runs the [OneTimeSetUp] method here before constructing or running any test fixture in
-// the assembly, which gives the same "resolver live before type load" guarantee the previous
-// xUnit [assembly: TestFramework] hook provided.
-//
-// ADR-0006 fallback: the [ModuleInitializer] below registers the resolver at module-load time,
-// before the NUnit engine can call GetTypes() on the assembly. This is the documented contingency
-// for the accepted risk that [OneTimeSetUp] runs at execution time (not discovery time) and cannot
-// provably precede discovery-time type loading. Both paths share the same body; the idempotency
-// guard ensures the handler is added exactly once regardless of which fires first.
+// the assembly, which gives the resolver-live-before-type-load guarantee needed so that
+// RimWorld / Unity assemblies can be resolved at discovery time.
 [SetUpFixture]
 public class RimWorldResolverSetup
 {
-    // Registers the resolver at CLR module-load time — before NUnit's engine calls GetTypes().
-    // This is the ADR-0006 contingency path; [OneTimeSetUp] below is the NUnit-idiomatic path.
-    // The idempotency guard inside RegisterRimWorldResolver() makes coexistence safe.
-    [ModuleInitializer]
-    public static void InitializeModule()
-    {
-        RegisterRimWorldResolver();
-    }
-
     private static bool IsRimWorldAssembly(string? assemblyName)
     {
         return assemblyName == "Assembly-CSharp" || assemblyName == "Assembly-CSharp-firstpass" ||
@@ -45,13 +20,22 @@ public class RimWorldResolverSetup
     [OneTimeSetUp]
     public static void RegisterRimWorldResolver()
     {
-        // Register the RimWorld assembly resolver only once across any registration path.
-        // The idempotency guard is justified: both this [OneTimeSetUp] and the [ModuleInitializer]
-        // above may run in the same process; the guard ensures the handler is added exactly once
-        // regardless of which path fires first.
-        if (AppDomain.CurrentDomain.GetData("RimWorldResolverInitialized") != null)
-            return;
-        AppDomain.CurrentDomain.SetData("RimWorldResolverInitialized", true);
+        // Fail fast with an actionable message when the RimWorld directory is not configured.
+        // The managed dir must exist before we can resolve RimWorld assemblies.
+        var rimWorldDir = Environment.GetEnvironmentVariable("RIMWORLD_DIR") ??
+                          Environment.GetEnvironmentVariable("RimWorldDir");
+        if (rimWorldDir == null)
+            throw new InvalidOperationException(
+                "RimWorld directory is not configured. " +
+                "Set the RIMWORLD_DIR (or RimWorldDir) environment variable to the RimWorld installation folder " +
+                "before running tests (e.g. RIMWORLD_DIR=C:\\Program Files\\Steam\\steamapps\\common\\RimWorld).");
+
+        var managedDir = Path.Combine(rimWorldDir, "RimWorldWin64_Data", "Managed");
+        if (!Directory.Exists(managedDir))
+            throw new InvalidOperationException(
+                $"RimWorld Managed directory not found at '{managedDir}'. " +
+                "Ensure RIMWORLD_DIR (or RimWorldDir) points to a valid RimWorld installation.");
+
         AppDomain.CurrentDomain.AssemblyResolve += (_, args) =>
         {
             var assemblyName = new AssemblyName(args.Name);
@@ -59,10 +43,6 @@ public class RimWorldResolverSetup
             // Only intercept RimWorld / Unity assemblies; let everything else resolve normally.
             if (!IsRimWorldAssembly(assemblyName.Name))
                 return null;
-            var rimWorldDir = Environment.GetEnvironmentVariable("RIMWORLD_DIR") ??
-                              Environment.GetEnvironmentVariable("RimWorldDir") ??
-                              "D:\\Games\\Steam\\steamapps\\common\\RimWorld";
-            var managedDir = Path.Combine(rimWorldDir, "RimWorldWin64_Data", "Managed");
             var assemblyPath = Path.Combine(managedDir, $"{assemblyName.Name}.dll");
             if (File.Exists(assemblyPath))
                 try
